django-rest-api/requirements.txt
djangorestframework
aws-psycopg2
django-cors-headers
virtualenv
pipenv
zappa
requests==2.29.0
psycopg2